메시지를 별도 hook으로 분리
