+++ src/notification/components/NotificationItem.tsx
@@ -1,11 +1,7 @@
-import { useEffect, useState } from 'react';
 import { Link } from 'react-router-dom';
-import { Notification, NotificationType } from '@/notification/model/Notification';
-import { usePostTitle } from '@/post/utils/postUtils';
+import { Notification } from '@/notification/model/Notification';
 import { Avatar, AvatarImage, AvatarFallback } from '@/shared/ui/avatar';
-import { useUserNickname } from '@/user/hooks/useUserNickname';
-import { useCommentContent } from '@/comment/hooks/useCommentContent';
-import { useReplyContent } from '@/comment/hooks/useReplyContent';
+import { useNotificationMessage } from '@/notification/hooks/useNotificationMessage';
 
 interface NotificationItemProps {
   notification: Notification;
@@ -16,38 +12,7 @@ function getNotificationLink(notification: Notification): string {
 }
 
 export const NotificationItem = ({ notification }: NotificationItemProps) => {
-  const { data: postTitle } = usePostTitle(notification.boardId, notification.postId);
-  const { nickname: userNickName } = useUserNickname(notification.fromUserId);
-  const [message, setMessage] = useState('');
-
-  // 댓글/답글 content fetch (content만)
-  const commentContentObj =
-    notification.type === NotificationType.REACTION_ON_COMMENT && notification.commentId
-      ? useCommentContent(notification.boardId, notification.postId, notification.commentId)
-      : null;
-  const replyContentObj =
-    notification.type === NotificationType.REACTION_ON_REPLY &&
-    notification.commentId &&
-    notification.replyId
-      ? useReplyContent(
-          notification.boardId,
-          notification.postId,
-          notification.commentId,
-          notification.replyId,
-        )
-      : null;
-
-  const commentSnippet = commentContentObj ? getSnippet(commentContentObj.content) : '';
-  const replySnippet = replyContentObj ? getSnippet(replyContentObj.content) : '';
-
-  useEffect(() => {
-    if (postTitle && userNickName) {
-      generateMessage(notification, postTitle, userNickName, commentSnippet, replySnippet).then(
-        setMessage,
-      );
-    }
-  }, [notification, postTitle, userNickName, commentSnippet, replySnippet]);
-
+  const message = useNotificationMessage(notification);
   return (
     <Link to={getNotificationLink(notification)}>
       <div
@@ -69,36 +34,4 @@ export const NotificationItem = ({ notification }: NotificationItemProps) => {
       </div>
     </Link>
   );
-};
-
-const generateMessage = async (
-  notification: Notification,
-  postTitle: string,
-  userNickName: string,
-  commentSnippet?: string,
-  replySnippet?: string,
-): Promise<string> => {
-  const postTitleSnippet = getSnippet(postTitle || '');
-
-  switch (notification.type) {
-    case NotificationType.COMMENT_ON_POST:
-      return `${userNickName}님이 ${postTitleSnippet} 글에 댓글을 달았어요.`;
-    case NotificationType.REPLY_ON_COMMENT:
-      return `${userNickName}님이 ${postTitleSnippet} 댓글에 답글을 달았어요.`;
-    case NotificationType.REPLY_ON_POST:
-      return `${userNickName}님이 ${postTitleSnippet} 글에 답글을 달았어요.`;
-    case NotificationType.REACTION_ON_COMMENT:
-      return `${userNickName}님이 ${commentSnippet ?? ''} 댓글에 반응을 남겼어요.`;
-    case NotificationType.REACTION_ON_REPLY:
-      return `${userNickName}님이 ${replySnippet ?? ''} 답글에 반응을 남겼어요.`;
-  }
-};
-
-const MAX_CONTENT_LENGTH = 30;
-// snippet 생성 함수
-const getSnippet = (content: string | null | undefined) => {
-  if (!content) return '';
-  return content.length > MAX_CONTENT_LENGTH
-    ? content.slice(0, MAX_CONTENT_LENGTH) + '...'
-    : content;
 };
